Moved some utility functions out of main msm_we
